refactor!: вынесение моделей из папки v2
