+++ anilibria/api/models/__init__.py
@@ -1,3 +1,5 @@
-from .v2 import *
+from .title import *
+from .misc import *
+from .enums import *
 from .attrs_utils import *
 from .cattrs_hooks import *
