--- anilibria/api/models/v2/__init__.py
@@ -1,3 +1,0 @@
-from .title import *
-from .misc import *
-from .enums import *
